+++ access-checkout/src/test/java/com/worldpay/access/checkout/testutils/AccessEditTextUtils.kt
@@ -4,7 +4,7 @@ import com.worldpay.access.checkout.ui.AccessCheckoutEditText
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 
-fun createAccessEditTextMock(returnValue: String): AccessCheckoutEditText {
+fun createAccessCheckoutEditTextMock(returnValue: String): AccessCheckoutEditText {
     val accessCheckoutEditTextMock = mock<AccessCheckoutEditText>()
     whenever(accessCheckoutEditTextMock.text).thenReturn(returnValue)